+++ src/main/java/util/CalculatorUtil.java
@@ -1,14 +1,14 @@
 package util;
 
 public class CalculatorUtil {
-    private CalculatorUtil(){
+    private CalculatorUtil() {
     }
 
     public static int divideWholeNumber(int target, int divideNumber) {
         return (int) Math.round(divide(target, divideNumber));
     }
 
-    public static double divide(int target, int divideNumber) {
+    public static double divide(long target, long divideNumber) {
         if (divideNumber == 0) {
             return 0;
         }
